Add files via upload
Upload for the challenge on JS. I found the code for this on a website here https://hackr.io/blog/how-to-build-rock-paper-scissors-in-javascript. However, I made sure to comment everything and spent quite a long time studying the whole code to make sure I understood it properly. I didn't do the score part or anything past that as that is not necessary. If you count this as cheating, send me to prison.